server/requirements.txt
pipenv
django
mysqlclient
graphene-django
django-graphql-jwt